multiple_linear_regression
This example allowed me to understand the Multiple regression :  Multiple linear regression is a statistical technique that models the relationship between a dependent variable and two or more independent variables by fitting a linear equation to the observed data. It extends simple linear regression, which only involves one predictor, to multiple predictors.